requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
dnspython==2.5.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
pyyaml==6.0.1
python-json-logger==2.0.7
